user module

+++ src/main/java/com/canglong/controller/HomeController.java
@@ -18,7 +18,7 @@ public class HomeController {
 	private UserService userService;
 	
 	@RequestMapping(value="/user/{userId}", method=RequestMethod.GET)
-	public String hello(@PathVariable("userId") Integer id, Model model) {
+	public String hello(@PathVariable("userId") Long id, Model model) {
 		User user = userService.getUser(id);
 		model.addAttribute("user", user);
 		return "user";
